Added sprite index to packets. Added pack and unpack functions to netcode

+++ src/settings.h
@@ -6,7 +6,7 @@ typedef struct
 {
     char name[PLAYER_NAME_MAX+1];
     uint32_t color;
-    int sprite_index;
+    uint8_t sprite_index;
 } Settings;
 
 bool settings_save();
